+++ CHANGELOG.md
@@ -5,8 +5,12 @@
 - Re-implemented model debugger to rely on methods of the class [#1581](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1581)
 - Added grouping of the fields by the superclass [#1581](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1581)
 - Added support of the custom model methods [#1581](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1581)
+- Added type-aware custom icons for each value node [#1582](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1582)
 - Renderer is not always created first time [#1578](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1578)
 - Handle non-property fields, such as `PK` [#1580](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1580)
+
+### `Type System` enhancements
+- Introduced custom icon for `dynamic` attributes [#1582](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1582)
 
 ## [2025.2.4.1]